Add null props guard to VideoDetail
This component is depending on information being passed down from its
parent, which is originally fetched from Youtube's API. It is almost
certain that the component will render before that request has
completed.

In this case, we should return something that is not dependent on the
props argument being defined.