Set up React Router Switch
React router defaults to fuzzy matching of routes, and will allow
multiple matches to fire and render their components by default (this is
*not* how most web app routers work).

In order to get more 'normal' behavior, use the Switch component to wrap
the routes. The Switch will act on the first match it finds, instead of
all.

This means that the most specific routes must occur first:

good:

<Switch>
  <Route path="/posts/new" />
  <Route path="/posts" />
</Switch>

bad:

<Switch>
  <Route path="/posts" />
  <Route path="/posts/new" />
</Switch>