Link from index to new post form
using React Router Dom's `Link` component.

We don't use an anchor tag because we don't want the browser to issue a
new HTTP GET request--we just want to hit the router inside the app JS
bundle.

`react-router-dom` includes a `Link` component that's more or less a
functional drop in for anchor tags, but instead of issuing a GET
request, it makes a call to the router.

The `to=` attribute takes a path that exists as a route in the router.